Script for run in virtualenv
There are now two main bash scripts: run_docker.sh and run_virtualenv.sh

Location of directory with resources is passed to python as
an environmental variable, which is set in scripts mentioned above.

In Seq2Seq algorithm training on GPU is possible now
(if GPU is available and configured for tensorflow).